moved middleware to util
