+++ src/api/assignments.ts
@@ -1,13 +1,25 @@
 import { Router, Request, Response, NextFunction} from "express";
 import {collection} from "firebase/firestore";
 import {db} from "./firebaseConfig";
-import { requiredInBody } from "./middleware.js";
+import { requiredInBody } from "../util/middleware.js";
+import { requireAuthentication } from "../util/authentication.js";
 
 const router = Router();
 
-router.post('/assignments', requiredInBody(["courseID", "title", "points", "due"]), (req: Request, res: Response) => {
-    const assignmentsCollection = collection(db, 'Assignments');
+router.post('/assignments', requireAuthentication, requiredInBody(["courseID", "title", "points", "due"]), (req: Request, res: Response) => {
+    //const assignmentsCollection = collection(db, 'Assignments');
+    return res.status(200);
 
+});
+
+router.get('/assignments/:id', (req: Request, res: Response) => {
+
+    return res.status(200);
+});
+
+router.patch('/assignments/:id', (req: Request, res: Response) => {
+
+    return res.status(200);
 })
 
 export default router;
